Address datetime issues, carriage-returns in psql, and varchar vs. text column distinction.
- Add 'compress_varchar' option which will find the largest entry in the
column, and set the VARCHAR(length), where length = len(largest_entry)
- In Postgresql, carriage_returns and newline chars must be removed
from the string. This is because psycopg2.copy_from() function does
not allow you to set the "ENCLOSED BY '"' option, causing these chars
be interpreted as literal newlines => the COPY FROM cmd thinks these
mean the 'end of the row' when importing...This is a temporary solution,
and a performance concern. Correct solution is to patch psycopg2.
- Postgresql does not support DATETIMES, but instead has a TIMESTAMP
column type. Thus, datetime columns should be converted to TIMESTAMP
accordingly
- Dates w/ years prior to 1900 are not able to be converted by the
strftime() method in Python. A fix has been applied in this release
for all 5 supported databases.
- Varchar() cols should be converted to TEXT columns if they are greater
than 256 bytes.