current_ordered_table_colums[pk] index issue [Issue #6]
Because current_ordered_table_columns is a list, and because the
`pk` variable above is a **string**, we want to call:

`current_ordered_table_columns.index(pk)`

to get the index of pk (i.e. what number column `pk` is in the
given row).

(This is an error due to regression following performance updates,
where each column is stored as a list and not a dictionary).